requirements.txt
numpy~=1.19.0
sklearn~=0.0
scikit-learn~=0.23.1
pandas~=1.0.4
PyYAML~=5.3.1
scipy~=1.4.1
matplotlib~=3.2.1
xgboost~=1.1.1